成功复现3dgs的原工程，添加了一些注释，记载了常用工程

+++ tiny-nerf/Plenoxels.py
@@ -86,8 +86,8 @@ class PlenoxelModel(nn.Module):
         grid = pts_norm.view(1, -1, 1, 1, 3)  # [1, N_pts, 1, 1, 3]
         
         features = F.grid_sample(
-            voxel_features,
-            grid,
+            voxel_features.float(),
+            grid.float(),
             align_corners=True,
             mode='bilinear',
             padding_mode='border'
